[util/hb-subset] Parse --text & --unicodes
